Add check for delete item

+++ src/Overlord.js
@@ -39,16 +39,19 @@ export default class Overlord extends React.Component {
 
   handleClick(e) {
     console.log(e.item.id);
-    this._db
-      .collection("queue")
-      .doc(e.item.id)
-      .delete()
-      .then(function() {
-        console.log("Document successfully deleted!");
-      })
-      .catch(function(error) {
-        console.error("Error removing document: ", error);
-      });
+    const r = confirm("oh really?");
+    if (r == true) {
+      this._db
+        .collection("queue")
+        .doc(e.item.id)
+        .delete()
+        .then(function() {
+          console.log("Item successfully deleted!");
+        })
+        .catch(function(error) {
+          console.error("Error removing item: ", error);
+        });
+    }
   }
 
   render() {
